authorizes users base on their role, fix job service: HR can only manage job of their company

+++ src/main/java/com/bugboo/CareerConnect/service/JobService.java
@@ -3,6 +3,7 @@ package com.bugboo.CareerConnect.service;
 import com.bugboo.CareerConnect.domain.Company;
 import com.bugboo.CareerConnect.domain.Job;
 import com.bugboo.CareerConnect.domain.Skill;
+import com.bugboo.CareerConnect.domain.User;
 import com.bugboo.CareerConnect.domain.dto.request.job.RequestCreateJobDTO;
 import com.bugboo.CareerConnect.domain.dto.request.job.RequestUpdateJobDTO;
 import com.bugboo.CareerConnect.domain.dto.response.ResponsePagingResultDTO;
@@ -11,6 +12,7 @@ import com.bugboo.CareerConnect.repository.CompanyRepository;
 import com.bugboo.CareerConnect.repository.JobRepository;
 import com.bugboo.CareerConnect.repository.SkillRepository;
 import com.bugboo.CareerConnect.type.exception.AppException;
+import com.bugboo.CareerConnect.utils.JwtUtils;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -23,13 +25,14 @@ import java.util.List;
 public class JobService {
     private final JobRepository jobRepository;
     private final SkillRepository skillRepository;
-    private final CompanyRepository companyRepository;
+
     private final ApplicationEventPublisher eventPublisher;
-    public JobService(JobRepository jobRepository, SkillRepository skillRepository, CompanyRepository companyRepository, ApplicationEventPublisher eventPublisher) {
+    private final JwtUtils jwtUtils;
+    public JobService(JobRepository jobRepository, SkillRepository skillRepository, ApplicationEventPublisher eventPublisher, JwtUtils jwtUtils) {
         this.jobRepository = jobRepository;
         this.skillRepository = skillRepository;
-        this.companyRepository = companyRepository;
         this.eventPublisher = eventPublisher;
+        this.jwtUtils = jwtUtils;
     }
 
     public ResponsePagingResultDTO getAllJobs(Specification<Job> specification, Pageable pageable) {
@@ -42,11 +45,15 @@ public class JobService {
     }
 
     public Job createJob(RequestCreateJobDTO requestCreateJobDTO) {
-        Company company = companyRepository.findById(requestCreateJobDTO.getCompanyId())
-                .orElseThrow(() -> new AppException("Invalid company id",400));
+        User user = jwtUtils.getCurrentUserLogin();
+        if(user.getCompany() == null && user.getRole().getName().equals("ROLE_HR")){
+            throw new AppException("HR must have a company",400);
+        }
+
         List<Skill> skillList = skillRepository.findByIdIn(requestCreateJobDTO.getSkillIds());
         if (skillList.isEmpty())
             throw new AppException("Invalid skill id",400);
+
         Job job = new Job();
         job.setName(requestCreateJobDTO.getName());
         job.setLocation(requestCreateJobDTO.getLocation());
@@ -55,7 +62,7 @@ public class JobService {
         job.setLevel(requestCreateJobDTO.getLevel());
         job.setDescription(requestCreateJobDTO.getDescription());
         job.setStartDate(requestCreateJobDTO.getStartDate());
-        job.setCompany(company);
+        job.setCompany(user.getCompany());
         job.setSkills(skillList);
         job = jobRepository.save(job);
         // notify subscribers
@@ -64,15 +71,23 @@ public class JobService {
         return job;
     }
 
+    boolean isJobBelongToHR(Job job){
+        return job.getCompany().getId() == jwtUtils.getCurrentUserLogin().getCompany().getId();
+    }
+
     public void deleteJob(int id) {
         Job job = jobRepository.findById(id).orElseThrow(() -> new AppException("Job not found with that id",400));
+        if(!isJobBelongToHR(job))
+            throw new AppException("You can't delete this job",403);
         jobRepository.delete(job);
     }
 
     public Job updateJob(RequestUpdateJobDTO requestUpdateJobDTO) {
         Job jobDB = jobRepository.findById(requestUpdateJobDTO.getId()).orElseThrow(() -> new AppException("Job not found with that id",400));
-        Company company = companyRepository.findById(requestUpdateJobDTO.getCompanyId())
-                .orElseThrow(() -> new AppException("Invalid company id",400));
+
+        if(!isJobBelongToHR(jobDB))
+            throw new AppException("You can't update this job",403);
+
         List<Skill> skillList = skillRepository.findByIdIn(requestUpdateJobDTO.getSkillIds());
         if(skillList.isEmpty())
             throw new AppException("Invalid skill id",400);
@@ -83,7 +98,6 @@ public class JobService {
         jobDB.setLevel(requestUpdateJobDTO.getLevel());
         jobDB.setDescription(requestUpdateJobDTO.getDescription());
         jobDB.setStartDate(requestUpdateJobDTO.getStartDate());
-        jobDB.setCompany(company);
         jobDB.setSkills(skillList);
         return jobRepository.save(jobDB);
     }
